refactor(sbom): Factor out sbom generation for artefact
Enable re-use, e.g. in BlackDuck extension, without introducing
dependency on crypto extension.
