+++ syft.py
@@ -1,9 +1,20 @@
 import enum
 import logging
+import os
 import subprocess
+import tarfile
+import tempfile
+
+import oci.client
+import ocm
+
+import dockerutil
+import secret_mgmt.aws
+import secret_mgmt.oci_registry
 
 
 logger = logging.getLogger(__name__)
+own_dir = os.path.abspath(os.path.dirname(__file__))
 
 
 class SyftSbomFormat(enum.StrEnum):
@@ -39,3 +50,94 @@ def run_syft(
         raise
 
     return sbom_raw
+
+
+def generate_raw_sbom_for_artefact(
+    component: ocm.Component,
+    access: ocm.Access,
+    secret_factory,
+    oci_client: oci.client.Client,
+    aws_secret_name: str | None = None,
+) -> str:
+    if access.type is ocm.AccessType.OCI_REGISTRY:
+        access: ocm.OciAccess
+
+        oci_secret = secret_mgmt.oci_registry.find_cfg(
+            secret_factory=secret_factory,
+            image_reference=access.imageReference,
+        )
+
+        if oci_secret:
+            dockerutil.prepare_docker_cfg(
+                image_reference=access.imageReference,
+                username=oci_secret.username,
+                password=oci_secret.password,
+            )
+
+        return run_syft(source=access.imageReference)
+
+    elif access.type is ocm.AccessType.S3:
+        access: ocm.S3Access
+
+        aws_secret = secret_mgmt.aws.find_cfg(
+            secret_factory=secret_factory,
+            secret_name=aws_secret_name,
+        )
+        s3_client = aws_secret.session.client('s3')
+
+        if isinstance(access, ocm.LegacyS3Access):
+            bucket = access.bucketName
+            key = access.objectKey
+        else:
+            bucket = access.bucket
+            key = access.key
+
+        fileobj = s3_client.get_object(Bucket=bucket, Key=key)['Body']
+
+        def tar_filter(member: tarfile.TarInfo, dest_path: str) -> tarfile.TarInfo | None:
+            if member.islnk() or member.issym():
+                if os.path.isabs(member.linkname):
+                    return None
+            return member
+
+        with tempfile.TemporaryDirectory(dir=own_dir) as tmp_dir:
+            s3_path = os.path.join(tmp_dir, 's3')
+
+            with tarfile.open(fileobj=fileobj, mode='r|*') as tar:
+                tar.extractall(
+                    path=s3_path,
+                    filter=tar_filter,
+                )
+
+            return run_syft(source=s3_path)
+
+    elif access.type is ocm.AccessType.LOCAL_BLOB:
+        access: ocm.LocalBlobAccess | ocm.LocalBlobGlobalAccess
+
+        if access.globalAccess:
+            image_reference = access.globalAccess.ref
+            digest = access.globalAccess.digest
+        else:
+            image_reference = component.current_ocm_repo.component_version_oci_ref(
+                name=component.name,
+                version=component.version,
+            )
+            digest = access.localReference
+
+        blob = oci_client.blob(
+            image_reference=image_reference,
+            digest=digest,
+            stream=True,
+        )
+
+        with tempfile.TemporaryDirectory(dir=own_dir) as tmp_dir:
+            local_blob_path = os.path.join(tmp_dir, 'local_blob')
+
+            with open(local_blob_path, 'wb') as file:
+                for chunk in blob.iter_content(chunk_size=4096):
+                    file.write(chunk)
+
+            return run_syft(source=local_blob_path)
+
+    else:
+        raise ValueError(f'dont know how to handle {access.type=}')
